+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Releases Changelog
 
+## 1.9.0
+- Added a write_schema field on db and api for dataset. This will be passed to the task as env var `WRITE_SCHEMA`.
+
+  The existing database `schema` value will be passed as `CDM_SCHEMA` to the task env var
+
 ## 1.8.0
 - **BREAKING CHANGE**: Migration from nginx-ingress to the gateway api based controller, traefik. This will need manual intervention. See the [gateway-api-installer.sh](./scripts/gateway-api-installer.sh) script to one-step setup. If not, the chart will let you know some components are missing and provide instructions.